Refactory: LinePath 관련 Steps를 따로 분리
