+++ src/test/java/nextstep/subway/acceptance/LineSteps.java
@@ -2,13 +2,14 @@ package nextstep.subway.acceptance;
 
 import static org.assertj.core.api.Assertions.*;
 
+import java.util.HashMap;
+import java.util.Map;
+
+import org.springframework.http.MediaType;
+
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import org.springframework.http.MediaType;
-
-import java.util.HashMap;
-import java.util.Map;
 
 public class LineSteps {
     public static ExtractableResponse<Response> 지하철_노선_생성_요청(String name, String color) {
@@ -84,18 +85,5 @@ public class LineSteps {
 
     public static void 지하철_노선에_역들을_확인(ExtractableResponse<Response> response, Long ...stationIds) {
         assertThat(response.jsonPath().getList("stations.id", Long.class)).containsExactly(stationIds);
-    }
-
-    public static ExtractableResponse<Response> 지하철_최단_경로_조회(Long sourceId, Long targetId) {
-        Map<String, String> params = new HashMap<>();
-        params.put("source", sourceId + "");
-        params.put("target", targetId + "");
-        return RestAssured
-            .given().log().all()
-            .queryParams(params)
-            .accept(MediaType.APPLICATION_JSON_VALUE)
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .when().get("/paths")
-            .then().log().all().extract();
     }
 }
